requirements.txt
requests==2.31.0
openai==1.35.10
httpx==0.25.2
python-dotenv==1.0.0
pytest==7.4.0
pytest-mock==3.12.0